producers/requirements.txt
kafka-python
pandas
websockets
